IRRE fields image and media on tt_content now use exact file count
Previously, image was set to a fixed value of 1. Upon saving however, BE changed this to the exact image count. Migration did not set a value for media before which meant "File Links" (also called "uploads") did not show any files on FE after migration and needed saving the content element once in BE to update the media count.

For now, only image & media are white-listed to be updated upon migration. If this is required for *any* IRRE field, just checking for TCA config may be sufficient (and necessary).