migraterelations now changes layout field on CType uploads
CType "uploads" (called "file links" on backend) will show different
on backend than when dam_filelinks was used. New default for
compatibility with dam_filelinks is to set layout = 1 on all
modified "uploads" content elements. This behaviour is configurable
by an optional parameter --uploads-layout and can be disabled by
setting it to "null".