image caption, title and alt text handling can be configured now
As previous attempts at correctly migrating caption, title and alt
texts revealed that migration cannot be done in one or two uniform
ways but requires individual configuration depending on TS setup of
a website we are now introducing "chains". Please refer to command
reference and MigrateRelationsChainOptions.rst for details on how
this works.