[BUGFIX] Check existing references with right data
With command dammigration:migraterelations the check for existing
relations in sys_file_reference is done with a wrong dataset.
Instead of using the sys_file.uid as uid_local, the old tx_dam.uid is
taken as reference. This results in wrong or not existing references
and leads to multiple references on calling the command multiple times.